Migrate to NextJs
