--- src/Dashboard/Pages/BuySell.tsx
@@ -1,7 +1,0 @@
-export default function BuySell() {
-  return (
-    <>
-      <h2>Buy and Sell</h2>
-    </>
-  );
-}
