--- src/Dashboard/Pages/History.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const History = () => {
-  return (
-    <>
-      <h1>History</h1>
-      <p>History below.</p>
-    </>
-  );
-};
-
-export default History;
