--- src/Dashboard/Pages/Home/Movements.tsx
@@ -1,10 +1,0 @@
-export default function Movements() {
-  return (
-    <>
-      <div style={{ display: "flex", flexDirection: "column" }}>
-        <h3>Your top and bottom movers</h3>
-        <p>...</p>
-      </div>
-    </>
-  );
-}
