--- src/Dashboard/Pages/Orders.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Orders = () => {
-  return (
-    <>
-      <h1>Orders</h1>
-      <p>Orders Below</p>
-    </>
-  );
-};
-
-export default Orders;
