--- src/Dashboard/Pages/Transfers.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-const Transfers = () => {
-  return (
-    <>
-      <h1>Transfers</h1>
-      <p>Transfers below.</p>
-    </>
-  );
-};
-
-export default Transfers;
