+++ src/app/page.module.css
@@ -1,0 +1,56 @@
+/*!* page.module.css *!*/
+/*.modal {*/
+/*  display: flex;*/
+/*  position: fixed;*/
+/*  top: 0;*/
+/*  left: 0;*/
+/*  width: 100%;*/
+/*  height: 100%;*/
+/*  z-index: 1;*/
+/*  background-color: rgba(0, 0, 0, 0.5);*/
+/*  backdrop-filter: blur(5px);*/
+/*  justify-content: center;*/
+/*  align-items: center;*/
+/*}*/
+
+/*.modal-content {*/
+/*  display: flex;*/
+/*  flex-direction: column;*/
+/*  padding: 20px;*/
+/*  background-color: white;*/
+/*  box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.2);*/
+/*  border-radius: 8px;*/
+/*  width: 400px;*/
+/*  gap: 20px;*/
+/*}*/
+
+/*.input-container {*/
+/*  display: flex;*/
+/*  flex-direction: column;*/
+/*  gap: 10px;*/
+/*}*/
+
+/*input {*/
+/*  padding: 10px;*/
+/*  font-size: 1rem;*/
+/*  border: none;*/
+/*  border-radius: 5px;*/
+/*  background-color: #ffffff;*/
+/*  width: 100%;*/
+/*  max-width: 300px;*/
+/*  outline: none;*/
+/*}*/
+
+/*.button-container {*/
+/*  display: flex;*/
+/*  justify-content: flex-end;*/
+/*}*/
+
+/*#login-button {*/
+/*  font-size: 1.2rem;*/
+/*  padding: 8px 8px;*/
+/*}*/
+
+/*#login-button:hover {*/
+/*  color: #38778b;*/
+/*}*/
