--- src/style.css
@@ -1,43 +1,0 @@
-body {
-        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-          'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-          sans-serif;
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-}
-
-button {
-        height: 20px;
-}
-
-h1 {
-        vertical-align: top;
-}
-
-#contentContainer {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        overflow: hidden;
-        margin-right: 250px;
-        padding: 16px;
-}
-
-header {
-        position: relative;
-        height: 60px;
-        display: flex;
-        align-items: center;
-        padding: 0 20px;
-        background-color: #f5f5f5;
-}
-
-.home {
-        display: flex;
-        flex-direction: row;
-        flex-grow: 1;
-        overflow: hidden;
-}
-
-
-
